+++ foxes/core/data_calc_model.py
@@ -2,11 +2,14 @@ import numpy as np
 import xarray as xr
 from abc import abstractmethod
 from dask.distributed import progress
+from dask.diagnostics import ProgressBar
 
 from .model import Model
 from .data import Data
+from foxes.utils.runners import DaskRunner
 import foxes.variables as FV
 import foxes.constants as FC
+
 
 
 class DataCalcModel(Model):
@@ -247,15 +250,11 @@ class DataCalcModel(Model):
 
         # reorganize results Dataset:
         results = (
-            results.assign_coords({FV.VARS: out_vars}).to_dataset(dim=FV.VARS).persist()
+            results.assign_coords({FV.VARS: out_vars}).to_dataset(dim=FV.VARS)
         )
 
-        # try to show progress bar:
-        if algo.verbosity > 0:
-            try:
-                progress(results)
-            except ValueError:
-                pass
+        if DaskRunner.is_distributed() and len(ProgressBar.active):
+            progress(results.persist())
 
         # update data by calculation results:
         return results.compute()