+++ foxes/input/farm_layout/add_from_csv.py
@@ -5,7 +5,7 @@ from foxes.core import Turbine
 
 def add_from_csv(
     farm,
-    file_path,
+    data_source,
     col_index=None,
     col_name=None,
     col_x="x",
@@ -30,8 +30,8 @@ def add_from_csv(
     ----------
     farm : foxes.WindFarm
         The wind farm
-    file_path : str
-        The input csv file
+    data_source : str or pandas.DataFrame
+        The input csv file or data source
     col_index : str, optional
         The index column, or None
     col_name : str, optional
@@ -65,9 +65,12 @@ def add_from_csv(
 
     """
 
-    if verbosity:
-        print("Reading file", file_path)
-    data = pd.read_csv(file_path, index_col=col_index)
+    if isinstance(data_source, pd.DataFrame):
+        data = data_source
+    else:
+        if verbosity:
+            print("Reading file", data_source)
+        data = pd.read_csv(data_source, index_col=col_index)
 
     tmodels = turbine_parameters.pop("turbine_models", [])
     H = turbine_parameters.pop("H", None)