+++ foxes/utils/runners.py
@@ -1,9 +1,9 @@
 from abc import abstractmethod, ABCMeta
 from copy import deepcopy
 import dask
-from dask.diagnostics import ProgressBar
 from dask.distributed import Client, LocalCluster
-
+from dask.distributed import get_client
+from dask.diagnostics import ProgressBar
 
 class Runner(metaclass=ABCMeta):
     """
@@ -169,6 +169,14 @@ class DaskRunner(Runner):
         ):
             self.scheduler = "distributed"
 
+    @classmethod
+    def is_distributed(cls):
+        try:
+            get_client()
+            return True
+        except ValueError:
+            return False
+    
     def initialize(self):
         """
         Initialize the runner
@@ -182,10 +190,6 @@ class DaskRunner(Runner):
 
             self.print(self._cluster)
             self.print(f"Dashboard: {self._client.dashboard_link}\n")
-
-        else:
-            self._config0 = deepcopy(dask.config.config)
-            dask.config.config["scheduler"] = self.scheduler
 
         super().initialize()
 
@@ -232,8 +236,7 @@ class DaskRunner(Runner):
             self.print("\n\nShutting down dask cluster")
             self._client.close()
             self._cluster.close()
-
-        else:
-            dask.config.config["scheduler"] = self._config0
+        
+        dask.config.refresh()
 
         super().finalize()